Kill the table lookup to determine if we have hit the end of the pattern...
If we had pattern "Hund" patend would point to 'd' and at table entries
table['d'] and table['D'] we put a one. This is wasting space that we don't
have on the target, kills any cacheline (table is big)... instead do a
compare... these statements are now gone from the shark profile.

this table had plenty of zero on

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@507 974c56cd-8151-0410-98c4-7bb2f43fa8c1