[fatfs] Remove ff.c as we only use tiny fat...
The difference between fat and tiny fat is that fat would
support more than one drive/partition. We do not not have
such a requirement and can go with tinyfat alone.

Update Makefile's, remove inclusion of ff.h, change S_MAX_SIZ
to hardcoded 512 (like from the ff.h). Remove unused fattime.c
stub as we are not needing it.

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@565 974c56cd-8151-0410-98c4-7bb2f43fa8c1