[kernel] Use the gcc attribute to delcare the interrupt handlers
According to C33Startup_E.pdf we should use __attribute__((interrupt_handler))
for the interrupts. gcc will emit extra code to save and restore the
state of several registers and inserts a reti. Saving and restoring the
%alr and %ahr is something we did not do before but is crucial for good
operations. ahr/alr is the Arithmetic Low/High Register. I looked at the
assembly code and the reti was added there.

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@721 974c56cd-8151-0410-98c4-7bb2f43fa8c1