[rendering] Move the WebKit patches to host-tools/rendering/patches
I'm quite happy with the Content -> Position extraction of the patches
and would like to use that for the content preparation. This will work
nicely for any SVG, HTML, Text content that needs a layout for any
kind of language and text flow no matter where the content is coming
from.

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@797 974c56cd-8151-0410-98c4-7bb2f43fa8c1