[rendering] Introduce 'simple_code.py' to compile/code a text run
simple_code.py will take a render_text.blib and conver it to a
binary representation that can be used on the device to provide
text display. As the name says it. It is a really simple code that
will not be the final one (because wikipedia would not fit onto the
SD card with this one).

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@800 974c56cd-8151-0410-98c4-7bb2f43fa8c1