[guilib] Move some null checks out of the fast paths
Check if the font is valid in render_string, remove the
check from get_glyph. When the font was checked and the
glyph number is in range we can just assume get_glyph
is returning something valid and do not need to check stuff
in render_glyph. The size check in set_pixel needs to stay
for now.

Partially discussed with Daniel

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@807 974c56cd-8151-0410-98c4-7bb2f43fa8c1