[locate] Allow to save and restore any arbitary state of the lookup
The main use case is that a search for 'Torv' already lists 'Torvalds'
and that searching for 'Torva' would not lead any results anymore. So
the code is remembering the first hit of the search and when a new char
is appended it restarts the search from the first hit.

This can now be easily extended to provide a stack for scrolling or
a stack for fast forward/backward...

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@827 974c56cd-8151-0410-98c4-7bb2f43fa8c1