[lookup] The trigram database is more than the c33 can handle...
for some reasons the data of the table is too much for the
c33. It is dieing somewhere... even if everything should
just be in ram... probably some bad alignment... who knows.

We are now using malloc to get pages for the trigram array. This
is not immediately crashing the device...

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@857 974c56cd-8151-0410-98c4-7bb2f43fa8c1