[article] Move the ad-hoc rendering code to article...
the interface is faril simple. Open an article (do the error
checking and write a message), ask to paint a given page, open
another article, close the existing, repeat.

The implementation is anything but final, the coding itself
will change, the scrolling is not fast. Specially going a page
up is slow as it will be read from start again..

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@861 974c56cd-8151-0410-98c4-7bb2f43fa8c1