[rendering] Most sucky implementation for scrolling....
For page up and page down go back to the start and reread
the file and find the right page. This is the most sucky implementation
but it is easy and it is working. It is not meant to be final in
any way...

Scrolling is now in o(number of pages). The simple code is
not properly sorting the text runs making it hard to skip early..
e.g the break from the inner for loop is only breaking the inner
for loop and not everything... this needs to be fixed

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@865 974c56cd-8151-0410-98c4-7bb2f43fa8c1