Tidy up a few warnings and some Makefiles problems
Eliminate all the warnings so the the kernel can be compiled
with -Werror - nothing major, missing prototypes and const.

Fix fatfs Makefile not compiling all files because the common/Makefile
was included at the start - should be at the end.

Add message to common/Makefile so an error message is output
if it is included to early.  Also generate .asm33 files from
the compiler so we can inspect the generated code.

Split the cleaning in the master Makefile into:
clean
clean-toolchain
and a complete-clean to do both


git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@987 974c56cd-8151-0410-98c4-7bb2f43fa8c1