New serial output routine and message routine
This required fixing types.h - which only exists
as a compatibility include for common/lcd.c
since lcd is shared with the bootloader and the
bootloader does not have access to wikilib.h.

misc.[ch] are also for lcd.c compatibility.

wikilib now contains all of the kernel/types.h
items. 


git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@988 974c56cd-8151-0410-98c4-7bb2f43fa8c1