[gui-lib] Change the other fast path to work with img->width == WIDTH
Currently the fast path assumed that the image has the width of
the scanline. Change the code to assume that the image is covering
the full width of the screen and then memcpy line by line. It should
be faster than the set_pixel path.

The keyboard image is now showing up on the display and the set_pixel
and the 2nd fast path were tested.

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1089 974c56cd-8151-0410-98c4-7bb2f43fa8c1