[forth] Conversion to indirect threading model
The original model was direct threaded, but this is inconvienient as
it requires interfacing with machine code for the docolon routine.

Change to indirect threading meand that a colon definition only need a
pointer to the docolon routine, but requires an extra header cell to
give the location of the parameter area.

Signed-off-by: Christopher Hall <hsw@openmoko.com>


git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1120 974c56cd-8151-0410-98c4-7bb2f43fa8c1