[forth] Change the registers and sections
Change to lower registers, the C33 architecture allows pushing several
lower registers more efficiently than pushing several higher
registers.  (i.e. the pushn rX pushed r0..rX inclusive)

The C runtime always preserves R0..R3 if it should need to use them.

Therefore use R0..R3 for the inner interpreter.

Because of the way the C runtime is set up R4..R9 can be used by any
of the forth routines without needing to preserve them.

Use specific section names so the the forth dictionary and parameter
areas so that this data/code can be moved im memory separately from
the C .text/.data sections.

Debug was factored out because it was tripping some bug in the linker
as did not get the correct address; having it as an external symbol
made it link correctly.

sio register changes to be compatible and these routines are now C
callable.

This allow the forth system code to be compatible with C and be able
to call routines coded in C.

Signed-off-by: Christopher Hall <hsw@openmoko.com>


git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1133 974c56cd-8151-0410-98c4-7bb2f43fa8c1