[forth] Add the ability to load forth code from a file
This is not completely functional, as it can only deal with one open
file.  i.e. the source-id variable is not preserved.  (also the fileid
need to be adjusted, according to the standard the first fileid should
be 1 - presently zero is used)

Remove the LED code so this can be used for testing in an external
file.

Signed-off-by: Christopher Hall <hsw@openmoko.com>


git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1149 974c56cd-8151-0410-98c4-7bb2f43fa8c1