[forth] Replace regs.inc by c33regs.inc
This is designed to be more flexible and uses a macro rather than just
having simple equates.

It should be possible for forth to read this file so that all of the
I/O register can be forth words.

The new version was generated from the EPSON datasheet and therefore
should have the correct addresses.

Signed-off-by: Christopher Hall <hsw@openmoko.com>


git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1154 974c56cd-8151-0410-98c4-7bb2f43fa8c1