[bootloader] Intitial try for SDRAM configuration
This does not work, the SDRAM does not read properly
In one test, writing a word seemed to work, i.e. it readback correctly
sometimes.

Got this aabbccdd  as the word, but mostly read: 00dd00dd
it was possible to increment the byte dd sucessfully
only the first word was accessed.

Signed-off-by: Christopher Hall <hsw@openmoko.com>


git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1190 974c56cd-8151-0410-98c4-7bb2f43fa8c1