[bootloader] Multistage boot menu system
Added as separate system soth the original eeprom 1,2 system is not
disturbed.  It is important not to disturb the original system.

The mbr (first 512 bytes derived from eeprom-1st) can load code from
any 8kByte boundary (at offset 0x300 in the 8kByte block).

Initially the segment at 0x300 will be loaded - this will be menu
which at present scans for heade info in the 0x0..0x2ff area of each
8kByte block and displays a menu.  It then return a code that allows
the mbr to load another block.

So far there is hello world, a key test program, file loaders for
forth and kernel, an a 32MByte memory test.

This system should help the tester to swap out these examples for
small test programs to verify the operation of the circuit until SDRAM
and SDCARD interfaces have been debugged.

This boot loader system works on the eval board sucessfully and is
being used to test the SAMO A1 board at the moment.

Signed-off-by: Christopher Hall <hsw@openmoko.com>


git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1226 974c56cd-8151-0410-98c4-7bb2f43fa8c1