[rendering] The current approach does not yield a good result...
This is the path to canossa, instead of sinking more time into
the current huffman code approach to introduce dictionaries and
group more things together we will opt for another simple solution.

Currently it is plain text, with a special marker for a chain of
glyphs, x- and y-position and font changes.

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1229 974c56cd-8151-0410-98c4-7bb2f43fa8c1