[bootloader] Attempt to split up mbr and bootloader
From what I can tell the two bootloaders have only little
in common there is little reason to share a common directory,
maybe mbr is going to take over, maybe not... we will see but
by separating these two sets of things, things remain clean
and understandable.

The mbr Makefile is currently not perfoming as it should, I will
attempt to fix that. For now I'm adding the apps manually to
the list of targets.

This commit can be easily reverted if it conflicts with any plan
I'm not aware of.

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1233 974c56cd-8151-0410-98c4-7bb2f43fa8c1