[spacing] Compress the files as they can get pretty big
We open the file with append only and will have many many many
duplicates. Compress with gzip which should save us from the bad
and after it is done we can use the uniq to sort this out for
us. The alternative is to use any other kind of db...

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1246 974c56cd-8151-0410-98c4-7bb2f43fa8c1