[fatfs] Recalibrate delays in micro seconds
Also added extra delay.

This appears to make writing the th SD Card possible

The effect of this change is to slow down the wait ready
by an order of magnitude.  Possibly the card was
still performing write and the wait timed out too fast.

The mmc driver needs more work.

This is just a quick fix to make it usable.

Signed-off-by: Christopher Hall <hsw@openmoko.com>


git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1300 974c56cd-8151-0410-98c4-7bb2f43fa8c1