[glyphs] Start mapping glyphs ranges to [0:number of glyphs]
For chinese we don't want to create an array with 20.000 entries
just because the first 10.000 are empty and we happened to use
the glyph number 34234. Map all glyphs from their glyph id to
a smaller monotonely increasing number.

Special case font zero even more, for font zero we will have a
direct non altering mapping. If this wouldn't be the case gui-lib
would require a lookup table dependant on the result of this
script run.

Text that is formatted on the device does not require any change,
the coding will need to use the same mapping.

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1346 974c56cd-8151-0410-98c4-7bb2f43fa8c1