[gui-lib] Paint the image as it is...
This is an old regression from the time John introduced inverted
painting. When painting a image it should be drawn the way the pixels
are... and should be not interpreted in any way. Do this by introducing
another version of set_pixel without the #ifdef and share the
implementation with a #define.

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1380 974c56cd-8151-0410-98c4-7bb2f43fa8c1