[glyphs] Implement gen_spacing_hints using the new spacing-file
Update gen_spacing_hints to use the spacing-file as input for
the spacing between two glyphs. The mapping is not sorted in any
way and the implementation that needs to use this is forced to
do a linear search over this. This looks like a place for a good
optimisation...

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1386 974c56cd-8151-0410-98c4-7bb2f43fa8c1