[scripts] Add script that generates the indexfile from a database
We will have to combine the titles with the actual file and
offset. The best way I could think of was to use a sqlite3 instance
for doing the select. This script is the result... further changes
to generate the database and use the generated result will need
to be made.

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1387 974c56cd-8151-0410-98c4-7bb2f43fa8c1