[wiki-tools] Write the index title and hash to a database
Create a sqlite3 database for the title and hash, fill it
within parsingFinished, create the tables and index in the
constructor. This database has two tables one for Index data
and one for the file offsets the recently added extract_index.sh
is operating on this database schema. The wiki-tools is still
running while doing this change, there might be a bugfix
coming.

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1390 974c56cd-8151-0410-98c4-7bb2f43fa8c1