[wikilib] Do not compile image viewing by default.
The minimalistic and simple image viewer was added in order
to allow Will to easily view a image, e.g. to test different
fonts without going through the whole chain. There is no image
viewer in the spec, so only enable the development image viewer
when we want to. Converge torwards the real product.

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1391 974c56cd-8151-0410-98c4-7bb2f43fa8c1