[keyboard] autogenerate keyboard include file
The newly added kbd_to_include.py generates the keyboard_image.h whenever needed.

What was wrong before:
The conversion procedure was not reproducible as it required manual fiddling with file format conversions / copy & paste of text sections / wrong scripts names / inverting the file somehow ?? Furthermore, generated files are not to be added to the repository. The file (in its entity) should be generated (not just parts of it) as part of the normal build process.

Regressions:
Toppers does not compile anymore, hence I removed it from the default makefile target. Feel free to fix it.


git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1394 974c56cd-8151-0410-98c4-7bb2f43fa8c1