[common] add more lcd controls
This is not the correct place, but the other LCD controls
are here.

Note that not all controls apply to all platforms, the LCD need
refactoring.  Other LCD values are forund in common/lcd.h LCD_xxx overlap
FRAMEBUFFER_xxx in gui-lib/guilib.h givig a total of 5 values.

Each platform ought to define only 4 values:
  VRAM     width and height  (320x240) fixed by LCD buffers
  DISPLAY  width and height  (240x208) fixed by LCD glass

Signed-off-by: Christopher Hall <hsw@openmoko.com>


git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1409 974c56cd-8151-0410-98c4-7bb2f43fa8c1