[wiki-tools] Remove paranoia check to pass unit tests
Real world wikipedia has a mulitple line redirect. This is
unfortunate as this might lead to treating proper articles
as redirect when they happen to start with the magic words. In
a better world I would look up the spec and see... I should
find some time to read the php parser again...

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1431 974c56cd-8151-0410-98c4-7bb2f43fa8c1