[wiki-lib] Sketch out parser state machine...
It starts with

fNUMBER, Indicating which font to use.
x_y,     Indicating the relative x and y change
glyph(-glyph)* A list of glyphs to draw at this position. The spacing
         must be extracted using the fontfile.h glyph_spacing...
         This can be followed by 'f' to change the font, or a ',' to
         indicate a new x,y spacing...

I would just consume a byte at a time and convert it to a number...
using the number routine as in the title and offset routine (extract
this)...

git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1460 974c56cd-8151-0410-98c4-7bb2f43fa8c1