[jsp] Linker error (unbreak with assembler code)
There is a proble with the datapointer when linking the program.  The
linker is getting a value for a variable lower than the datapointer
value.  I have seen this before and fixed it by making the local
variables into register variables.

In this case adding register did not fix it, maybe becase it is a code
pointer.  Therfore I translated the code block to asdembler and added
a warning.

Basically this routine is testing to see if an external atexit
function is defined, I guess that it is not, which is why the linker
error occurs.

I added a #warning as a reminder to review this later.

Signed-off-by: Christopher Hall <hsw@openmoko.com>


git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1492 974c56cd-8151-0410-98c4-7bb2f43fa8c1