[common] optimise ADC routine
Make sure that the division in the final return is a power of 2, so
that the compiler can change to a shift.  This is to overcome a gcc
compiler bug that incldes the divide routine even if the inline
routine is never used.

Addition of this divide caused mbr to become too large.

Signed-off-by: Christopher Hall <hsw@openmoko.com>


git-svn-id: https://wikipediardware.googlecode.com/svn/trunk@1503 974c56cd-8151-0410-98c4-7bb2f43fa8c1