[drivers] remove cache call from mmc.c
The original cache code could no be used (broken) so remove the calls,
as these were usesless.

The original cache did not spedd multisector read and was only for
single sector read.  The algortithm was slow so most likely would have
slowed things down.  (also the malloc for the buffers was missing)

Signed-off-by: Christopher Hall <hsw@openmoko.com>