[00run.sh] fixes for time not found
Use bash since other shells (e.g. dash) may not have the time
command as a built-in and the system may not have the time
command in /usr/bin.

"eval" the debug lines so that the empty ${debug} will not skip
the search for built-in bash commands.

Also remove the time commands from the Makefile as these appear to
use the /usr/bin/time and not the bash built-in.

Signed-off-by: Christopher Hall <hsw@openmoko.com>