[grifo] refactoring low level driver code
This is a test program to check on the possibility of moving
all the driver code to one file and have the application in
another.

This program illustates this by having kernel.elf contain a few
drivers (serial port, timer, delay) and a sample.elf that is loaded
second to use these via a simple syscall interface.

Signed-off-by: Christopher Hall <hsw@openmoko.com>