[grifo] memory allocator
Heap memory is determined after loading the elf file by taking the
hishest unused address and a fixed stack depth then rounding these to
page sizes to initialise the allocator.

The allocator consolidate freespace on allocation.

The Allocation header allows for freespace on the before actual header
and before the start of data to allow for off by 1 (or -1 errors)

Signed-off-by: Christopher Hall <hsw@openmoko.com>