[grifo] add icons for each sample program
And greate a simple script to construct the .ico file from a .xpm
file.

The script has hard coded assumptions about the xpm format
i.e. '.' is black and space is white and there are only these two colours.
xpm is used since it is a text based format and can easily be changed in
a text editor.

Update Makefiles and ignore files to account for these changes

Signed-off-by: Christopher Hall <hsw@openmoko.com>