[imagetool] a new more flexible image encoder
Use getopt so that additional functions can be added in particular
inverted images can be created.

The bastic funtion is to treat black [rgb == (0,0,0)] as one state and
all other colours as another.  These two state give the two bitmap
values [0, 1] or [1, 0] if inverted.

Add rules to make rules for png -> h conversion.

Signed-off-by: Christopher Hall <hsw@openmoko.com>