[qt4-simulator] changes to run on 64 bit OS
Several of the structure used contain long and int fields, these
become the wrong size when compiled on 64 bit platform; so change to
intXX_t types to ensure the correct size.

There maybe more fields that are wrong, but this patch set allows the
program to run, but it will SEGV on scrolling some times.  Not sure if
this is a bug in field sizes or an actual race condition in handling
the mouse and key queue in the qt4 framework.  I suspect the latter.

Added sim4 target i.e.: make DESTDIR=image sim4
This runs the simulator on the current data files.

Also added sim4d to use gdb to help with finding the SEGV mentioned
above.

Signed-off-by: Christopher Hall <hsw@openmoko.com>