[offline-renderer]	Updated Mediwiki codebase
This commit updates mediwiki and fixes the template rendering
problems. All templates should parse correctly now. It also
removes a lot of old files that are not needed.

Note: Currently this only works for English Wikipedia as the
reference to templates are hardcoded in English. I will fix this
shortly.

Signed-off-by: Sean Moss-Pultz <sean@openmoko.com>