[makefiles] cleanup
Created UpdateMakefiles to simplify maintenance of the common code at
the top of each makefile.  The mastercopy of this is part of
definitions.mk and the UpdateMakefiles will copy a newer version to
each Makefile that requires it.

Also added macros to test for the presence of a executable program
e.g.

$(call REQUIRED_BINARY, sqlite3, sqlite3)

(Actually the detected program could be a script as long as it was
maked as executable)

Added a test for the presence of PHP class
e.g.

$(call REQUIRED_PHP, PDO, php5-sqlite)

These cause the make to fail is the requirements are not met and display
an appropriate apt-get command.

Signed-off-by: Christopher Hall <hsw@openmoko.com>