[mbr] update memory tests
Memory test to use the ram size routine.
Memory check to test if memory is mirrored

Add start of SDRAM to linker script, not able to use this
yet as the compiler restricts external addresses to 26 bits.
as the start, 0x10000000 is 29 bits this compiler trick:

  extern char memory_start;
  #define MEMORY_ADDRESS (&memory_start);

will alwways result in zero

Signed-off-by: Christopher Hall <hsw@openmoko.com>