[Makefiles] add merge operation
This is an optional step between parse and render.
it reduces the number of article html files to the value specified
by TOTAL_HTML_FILES on the make command line.

This allows for the possibliity of having more than 64 parsing
instances thne merging the generated files back to 64 or less
for the rendering stage.

Signed-off-by: Christopher Hall <hsw@openmoko.com>