[scripts,makefiles] change back to simple division, remove SQL, more..
Remove the time consuming SQL check to find the number of
articles to process, revert back to a simple equal sized division
of work for each process.

Add subdirectory to WORKDIR and TEMPDIR so files
are separated by language.

Make WIKI_DIR_SUFFIX?=pedia as the default subdirectory suffix

Put all temporary "stamp" files into the new work subdirectory

Update scripts to be compatible with the above

It should now be possible to process multiple languages without
having to clean a previous run of a different language.

Signed-off-by: Christopher Hall <hsw@openmoko.com>