[offline-renderer] fix nul padding when using newer Python
Older version of Python did not require explicit seek. this was likely
a bug in Python since the library reference says that truncate does not
change the file position.

On Lucid lack of the seek caused many nul bytes in all articles after the
first as the internal buffer file position was never reset to zero
prior to truncating it.

Signed-off-by: Christopher Hall <hsw@openmoko.com>