[offline-renderer] Fixes for too little vertical space
It is possible to get text outside of <p></p> which results in too
little space between the title and the text.  This was cause by
php5-tidy not being installed.

Ensuring the tidy module is there causes additional non-text
paragraphs to be inserted (e.g. <p><a ...></a></p>) that tidy does not
delete.  Modify the renderer to try to suppress these.

Signed-off-by: Christopher Hall <hsw@openmoko.com>