[offline-renderer] Uppercase the first character of template names
The template names now have their first character set to upper case
this is to be compatible with the PHP code which always uppercases the
first character of a title.

The Wikipedia data dumps have th first character uppercased, but
Wiktionary does not.  This change should improve the parsing of
Wiktionary dumps.

Signed-off-by: Christopher Hall <hsw@openmoko.com>