[scripts] fix negative time display on rendering analysis
With the addition of the count files it is no longer
necessary to initialise the rendering count from the parse count.
There are two outputs in the log file for the rendering count
the first is used to initialise the max value and the second
is used to indicate 100% completion.

Signed-off-by: Christopher Hall <hsw@openmoko.com>