[wiki,grifo] temperature display
Grifo ADC added channel to read temperature in centi-celcius

Wiki conditionally compiled code to display temperature on search screen
when the search string is empty.

set TEMPERATURE_DISPLAY=YES when compiling wiki

If compiled in, then a small 5 pixel high/wide 'T' shows up in the top
right to let you know something is there to press. when you press it it
cycles between C, F, and then back to the T icon.

Signed-off-by: Christopher Hall <hsw@openmoko.com>